fix(web): server-render topbar/sidebar nav links for SEO indexing
GSC reported 15 algorithm/explainer pages as "Discovered - currently
not indexed". The topbar and algorithm sidebar nav were injected
entirely client-side (el.innerHTML), so Googlebot's initial HTML
fetch saw zero internal links to these pages — they were reachable
only via sitemap.xml, reading as orphan pages.

Extract the render logic into src/nav-html.mjs, shared by the
browser bundle and a new Vite transformIndexHtml plugin that bakes
real <a href> links into every HTML entry at build time. Also adds
a static "Explore the algorithms" section to the homepage.
