+++ teeline-web/index.html
@@ -364,6 +364,12 @@
       <!-- /#step-content -->
     </main>
 
+    <section class="container" aria-labelledby="algorithms-index-heading">
+      <h2 id="algorithms-index-heading">Explore the algorithms</h2>
+      <p>Read how each solver works, with pseudocode and references:</p>
+      <div id="algorithms-index"></div>
+    </section>
+
     <footer id="app-footer" class="container">
       <span id="wasm-status" class="status-dot status-dot--loading"></span>
       <span id="wasm-version">Connecting…</span>
