+++ teeline-web/src/algorithms-sidebar.ts
@@ -1,27 +1,6 @@
-import { SOLVER_GROUPS, SOLVER_META, PAGED_SOLVERS } from './solver-groups'
+import { renderSidebarHtml } from './nav-html.mjs'
 
-export function renderSidebarHtml(currentId: string | null): string {
-  const groups = SOLVER_GROUPS.map(g => {
-    const items = g.ids.map(id => {
-      const meta = SOLVER_META[id]
-      if (!meta) return ''
-      const isCurrent = id === currentId
-      const inner = isCurrent
-        ? `<span aria-current="page">${meta.name}</span>`
-        : PAGED_SOLVERS.has(id)
-          ? `<a href="/algorithms/${id}/">${meta.name}</a>`
-          : `<span>${meta.name}</span>`
-      const currentClass = isCurrent ? ' sidebar-item--current' : ''
-      return `<li class="sidebar-item${currentClass}">${inner}</li>`
-    }).join('')
-    return `
-      <li class="sidebar-group">
-        <span class="sidebar-group-label">${g.label}</span>
-        <ul>${items}</ul>
-      </li>`
-  }).join('')
-  return `<ul>${groups}</ul>`
-}
+export { renderSidebarHtml }
 
 export function initSidebar(containerId = 'algo-sidebar'): void {
   const el = document.getElementById(containerId)
